requirements.txt
streamlit
google-genai
plotly
trafilatura